fix RMSNorm dtype: cast scale to input dtype for FA3 compatibility
The fp32 scale parameter promoted bf16 q/k tensors to fp32, causing
FlashAttention 3 to reject them. Cast scale to match input dtype.

+++ nanochat/gpt.py
@@ -50,7 +50,7 @@ class RMSNorm(nn.Module):
         self.scale = nn.Parameter(torch.ones(dim))
 
     def forward(self, x):
-        return F.rms_norm(x, (x.size(-1),)) * self.scale
+        return F.rms_norm(x, (x.size(-1),)) * self.scale.to(x.dtype)
 
 
 def has_ve(layer_idx, n_layer):
